:building_construction: Refactor project (#3)
* :building_construction: Refactor project

* :green_heart: Fix Action file
