+++ .github/workflows/go.yml
@@ -21,7 +21,7 @@ jobs:
           go-version: '1.20'
 
       - name: Build
-        run: go build -v ./...
+        run: go build -o ./GinDemo ./main.go
 
       - name: Run built binary & Test
         run: |
@@ -29,7 +29,7 @@ jobs:
 
           sleep 5
 
-          response=$(curl -s -o /dev/null -w "%{http_code}" http://localhost:8080/getTest)
+          response=$(curl -s -o /dev/null -w "%{http_code}" http://localhost:8080/apiDemo/getTest)
           if [ $response -eq 200 ]
           then
             echo "Test passed"
