+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "GinDemo/bootstrap"
+
+func main() {
+	bootstrap.RunServer()
+}
