+++ model/Request.go
@@ -1,0 +1,6 @@
+package model
+
+type RequestsBody struct {
+	Name    string   `json:"name"`
+	BankNos []string `json:"bankNos"`
+}
